Kiera Zone Blog

+++ app/Http/Controllers/PostsController.php
@@ -3,6 +3,7 @@
 namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Storage;
 use App\Post;
 
 class PostsController extends Controller
@@ -12,6 +13,11 @@ class PostsController extends Controller
      *
      * @return \Illuminate\Http\Response
      */
+    public function __construct()
+    {
+        $this->middleware('auth',['except' => ['index','show']]);
+    }
+
     public function index()
     {
         # $posts = Post::all();
@@ -37,16 +43,34 @@ class PostsController extends Controller
      */
     public function store(Request $request)
     {
-        $this->validate($request,['title' => 'required','body' => 'required']);
+        $this->validate($request,['title' => 'required','body' => 'required','cover_image' => 'image|nullable|max:1999']);
         
+        # Cover Image Process
+        if($request->hasFile('cover_image'))
+        {
+            $fileNameWithExtension = $request->file('cover_image')->getClientOriginalName();
+            $fileName = pathinfo($fileNameWithExtension,PATHINFO_FILENAME);
+            $extension = $request->file('cover_image')->getClientOriginalExtension();
+            $fileNameToStore = $fileName."__".time().".".$extension;
+            $path = $request->file('cover_image')->storeAs('public/coverImages',$fileNameToStore);
+        }
+        else
+        {
+            $fileNameToStore = 'noimage.jpg';
+        }
+
         $post = new Post;
         $post->title = $request->input('title');
         $post->body =  $request->input('body');
-        $post->user_id = 1;
+        $post->user_id = auth()->user()->id;
+        $post->cover_image = $fileNameToStore;
         
         if($post->save())
         {
-            return redirect('post')->with('success','Sucess In Adding a New Post!');
+            if(!(auth()->check()))
+                return redirect('post')->with('success','Sucess In Adding a New Post!');
+            else 
+                return redirect('dashboard')->with('success','Sucess In Adding a New Post!');
         }
 
         return redirect('post/create')->with('error','Cannot Insert a New Post!');
@@ -71,7 +95,12 @@ class PostsController extends Controller
      */
     public function edit($id)
     {
-        return view('posts.edit')->with('post',Post::find($id));
+        $post = Post::find($id);
+
+        if(auth()->user()->id != $post->user_id)
+            return redirect('/post')->with('error','Error Unauthorized action');
+
+        return view('posts.edit')->with('post',$post);
     }
 
     /**
@@ -84,9 +113,28 @@ class PostsController extends Controller
     public function update(Request $request, $id)
     {
         $post = Post::find($id);
-        $this->validate($request,["title" => "required","body"  => "required"]);
+        $this->validate($request,["title" => "required","body"  => "required","cover_image" => "image|nullable|max:1999"]);
+        
+        if($request->hasFile('cover_image'))
+        {
+            $fileNameWithExtension = $request->file('cover_image')->getClientOriginalName();
+            $fileName = pathinfo($fileNameWithExtension,PATHINFO_FILENAME);
+            $extension = $request->file('cover_image')->getClientOriginalExtension();
+            $fileNameToStore = $fileName."__".time().".".$extension;
+
+            $path = $request->file('cover_image')->storeAs('public/coverImages/',$fileNameToStore);
+
+            if($post->cover_image != 'noimage.jpg')
+                Storage::delete('public/coverImages/'.$post->cover_image);
+        }
+        
         $post->title = $request->input('title');
         $post->body = $request->input('body');
+        
+        if($fileNameToStore != "")
+            $post->cover_image = $fileNameToStore;
+
+        
         
         if($post->save()) return redirect('post')->with('success','Success In Updating a Record');
     }
@@ -100,6 +148,12 @@ class PostsController extends Controller
     public function destroy($id)
     {
         $post = Post::find($id);
+        
+        if(auth()->user()->id != $post->user_id)
+            return redirect('/post')->with('error','Error Unauthorized action');
+     
+        Storage::delete('public/coverImages/'.$post->cover_image);
+
         if($post->delete()) return redirect('post')->with('success','POST REMOVED');
 
     }
